surface missing parsers + add test
