+++ js/engine/tests/index.test.js
@@ -1,0 +1,27 @@
+const { EngineFactory } = require("../dist/index.cjs");
+
+const enginePromise = EngineFactory("./dist/semgrep-engine.wasm");
+
+describe("lookupParserName", () => {
+  test("handles base case", async () => {
+    const engine = await enginePromise;
+    expect(engine.lookupLang("java")).toEqual("java");
+  });
+  test("handles language aliases", async () => {
+    const engine = await enginePromise;
+    expect(engine.lookupLang("py")).toEqual("python");
+  });
+  test("handles invalid languages", async () => {
+    const engine = await enginePromise;
+    expect(engine.lookupLang("fake-language")).toBeNull();
+  });
+  test("detects an uninitialized language", async () => {
+    const engine = await enginePromise;
+    engine.execute(
+      "python",
+      `${__dirname}/test-rule-python.json`,
+      `${__dirname}/test.py`
+    );
+    expect(engine.getMissingLanguages()).toEqual(["python"]);
+  });
+});
