+++ js/engine/tests/test.py
@@ -1,0 +1,1 @@
+print("test")
